Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
asyncio-mqtt==0.16.2
python-multipart==0.0.12
scrapegraph-py